+++ torchtrainer/metrics/binary_accuracy.py
@@ -9,7 +9,7 @@ class BinaryAccuracy(Metric):
         self.total_count = 0
 
     def __call__(self, y_pred, y_true):
-        y_pred_round = y_pred.round().long()
+        y_pred_round = y_pred.round()
         self.correct_count += y_pred_round.eq(y_true).float().sum().item()
         self.total_count += len(y_pred)
         accuracy = 100. * float(self.correct_count) / float(self.total_count)
